==> 01_dog_tracking_data_processing.Rmd <==
---
title: "3D_tracking_data_processing"
author: "Christoph Voelter"
date: "11/03/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
```


### Read all csv files in folder
```{r}

read_plus <- function(flnm) {
    read_csv(flnm) %>% 
        mutate(filename = flnm)
}

xdata <-
    list.files(path = "./data/detector2/",
               pattern = "*.csv", 
               full.names = T) %>% 
    map_df(~read_plus(.))%>%
  separate(filename,  into=c(NA,NA,NA,NA,NA,NA,NA,NA,"subject","condition", NA, "date", NA), sep = "_")%>% #add info from filenames
  select(-frame_number, -"_cameras", -frame_timestamp, -type)

summary(xdata)
xdata%>%filter(is.na(name))%>%group_by(subject, condition)%>%summarise(min(x))
levels(as.factor(xdata$subject))
levels(as.factor(xdata$name))

levels(as.factor(xdata$condition))

### Object and door locations
object.data<-read.csv(file="data/3D_tracking_object_locations_chairs_plates.csv")
object.center.data<-read.csv(file="data/3D_tracking_object_locations_chairs_plates_centers.csv")

chair_window_ia_xr<-object.center.data$x[object.center.data$id=="chair_window"]+1
chair_window_ia_xl<-object.center.data$x[object.center.data$id=="chair_window"]-1
chair_window_ia_yl<-object.center.data$y[object.center.data$id=="chair_window"]
chair_window_ia_yh<-object.center.data$y[object.center.data$id=="chair_window"]+1.4

chair_door_ia_xr<-object.center.data$x[object.center.data$id=="chair_door"]+1
chair_door_ia_xl<-object.center.data$x[object.center.data$id=="chair_door"]-1
chair_door_ia_yl<-object.center.data$y[object.center.data$id=="chair_door"]-1.4
chair_door_ia_yh<-object.center.data$y[object.center.data$id=="chair_door"]

door_l_ia_xr<-1.466
door_l_ia_xl<-2.466
door_l_ia_yl<-2.7
door_l_ia_yh<-3.7

door_r_ia_xr<-(-2.265)
door_r_ia_xl<-(-1.265)
door_r_ia_yl<-2.7
door_r_ia_yh<-3.7

demographics.data<-read.csv(file="data/dog_tracking_project_data.csv")%>%
  mutate(subject=tolower(dog_name))%>%
  mutate(subject=fct_recode(subject, mira="mira3"))%>%
  mutate(condition=ifelse(condition_id=="pp","bp",ifelse(condition_id=="pa", "op", ifelse(condition_id=="ap", "sp", ifelse(condition_id=="aa", "np",NA)))))


```

```{r}
trial.duration.data <-  read.csv(file = "data/dog_exploration_3Dtracking_scoringdata.csv")%>%
  filter(Behaviour=="trial")%>%
  separate(Scoring,  into=c(NA,NA,NA,"subject","condition"), sep = "_") %>% #add info from filenames
  select(Start_Frame, Stop_Frame, subject, condition)


xx<-trial.duration.data%>%
  pivot_longer(cols=Start_Frame:Stop_Frame, values_to="frame_count", names_to="start_stop")%>%
  group_by(subject, condition)%>%
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))

```

### Filtered, interpolated, and smoothed data: all keypoints in one plot 

```{r}
xdata<-xdata%>%
  inner_join(xx)%>% #filter out data that don't belong to the trial
  mutate(subject_condition=paste(subject, condition, sep="_"))
subjects = unique(xdata$subject)
names = unique(xdata$name)
conditions =unique(xdata$condition)
subject_conditions =unique(xdata$subject_condition)
```


```{r}
library(zoo)
subject_dfs = list()
subject_average_dfs = list()
subject_plots = list()

for(subject_condition_ in subject_conditions){

subject_average_dfs[[subject_condition_]]<-xdata %>% filter(subject_condition == subject_condition_)%>% #calculate average coordinates across all keypoints
  filter(x<3.8 & x >(-3.5))%>% filter(z<1.5)%>%filter(y<3.7 & y>(-2.7))%>% #filter out unrealistic values
  group_by(frame_count)%>%
  summarise(average_x=mean(x),average_y=mean(y), average_z=mean(z))%>%
  select(frame_count, average_x, average_y, average_z)
  
subject_dfs[[subject_condition_]]<-xdata %>% filter(subject_condition == subject_condition_)%>%
  group_by(name)%>%
  inner_join(subject_average_dfs[[subject_condition_]], by="frame_count")%>% #add average values
  filter(x<3.8 & x >(-3.5))%>% filter(z<1.5)%>%filter(y<3.7 & y>(-2.7))%>% #filter out unrealistic values
    filter(abs(x-average_x)<1 & abs(y-average_y)<1 & abs(z-average_z)<1)%>% #filter out values that deviate from average value
  filter((abs(x-lag(x))<1.5 & abs(y-lag(y))<1.5 & abs(z-lag(z))<1.5 & abs(x-lead(x))<1.5 & abs(y-lead(y))<1.5 & abs(z-lead(z))<1.5))%>% #filter out jumps of 0.5m and more
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
  mutate(x_approx = na.approx(x,  na.rm = TRUE),y_approx = na.approx(y,  na.rm = TRUE), z_approx = na.approx(z,  na.rm = TRUE))%>% #first linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 20cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # second linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 20cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # third linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 20cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # 4th linear interpolation
  mutate(x_approx_roll = zoo::rollmean(x_approx, k = 3, fill = NA), y_approx_roll = zoo::rollmean(y_approx, k = 3, fill = NA),z_approx_roll = zoo::rollmean(z_approx, k = 3, fill = NA)) #rolling average
  
  subject_plots[[subject_condition_]] = 
    ggplot(subject_dfs[[subject_condition_]], aes(x=x_approx_roll, y=y_approx_roll)) +
    geom_path(alpha=0.2) + 
    geom_point(alpha=0.05) +
    annotate("rect", xmin=object.data$x[object.data$id=="dl1"], xmax=object.data$x[object.data$id=="dl2"], ymin=object.data$y[object.data$id=="dl1"],ymax=object.data$y[object.data$id=="dl2"], fill="dodgerblue", alpha=0.4)+
    annotate("rect", xmin=object.data$x[object.data$id=="dr1"], xmax=object.data$x[object.data$id=="dr2"], ymin=object.data$y[object.data$id=="dr1"],ymax=object.data$y[object.data$id=="dr2"], fill="dodgerblue", alpha=0.4)+
    annotate("rect", xmin=object.data$x[object.data$id=="wl1"], xmax=object.data$x[object.data$id=="wl2"], ymin=object.data$y[object.data$id=="wl1"],ymax=object.data$y[object.data$id=="wl2"], fill="dodgerblue", alpha=0.4)+
    annotate("rect", xmin=object.data$x[object.data$id=="wr1"], xmax=object.data$x[object.data$id=="wr2"], ymin=object.data$y[object.data$id=="wr1"],ymax=object.data$y[object.data$id=="wr2"], fill="dodgerblue", alpha=0.4)+
        annotate("rect", xmin=object.data$x[object.data$id=="door1"], xmax=object.data$x[object.data$id=="door2"], ymin=object.data$y[object.data$id=="door1"],ymax=object.data$y[object.data$id=="door2"], fill="darkorange", alpha=0.4)+
    annotate("rect", xmin=object.data$x[object.data$id=="window1"], xmax=object.data$x[object.data$id=="window2"], ymin=object.data$y[object.data$id=="window1"],ymax=object.data$y[object.data$id=="window2"], fill="darkorange", alpha=0.4)+
    theme(axis.text.x = element_text(angle = 90)) + 
    theme_bw()+ ggtitle(subject_condition_) + 
    facet_wrap(~name)+ #creates facets
    ylab("y") + xlab("x") + 
    xlim(-3.5, 3.8) + ylim(-2.7, 3.7) + 
    theme(plot.title = element_text(size = 24, face = "bold", hjust=0.5), axis.title.x = element_text(size = 16, face = "bold"), axis.title.y = element_text(size = 16, face = "bold"))
  
  print(subject_plots[[subject_condition_]])
  
  ggsave(subject_plots[[subject_condition_]], path="graphics/interpolated/", file=paste0("plot_facets_interpolated_", subject_condition_, ".pdf"), width = 40, height = 30, units = "cm", dpi=300)
  
}

```

### calculate distance

```{r}
library(zoo)

average_coordinates<-xdata %>%
  group_by(subject, condition, frame_count)%>%
  filter(x<3.8 & x >(-3.5))%>% filter(z<1.5)%>%filter(y<3.7 & y>(-2.7))%>% #filter out unrealistic values
  summarise(average_x=mean(x),average_y=mean(y), average_z=mean(z))%>%
  select(subject, condition, frame_count, average_x, average_y, average_z)

xdata.int<-xdata%>%
  group_by(subject, name, condition)%>%
  inner_join(average_coordinates)%>%
  filter(x<3.8 & x >(-3.5))%>% filter(z<1.5)%>%filter(y<3.7 & y>(-2.7))%>% #filter out unrealistic values
  filter(abs(x-average_x)<1 & abs(y-average_y)<1 & abs(z-average_z)<1)%>% #filter out values that deviate from average value
  filter((abs(x-lag(x))<1.5 & abs(y-lag(y))<1.5 & abs(z-lag(z))<1.5 & abs(x-lead(x))<1.5 & abs(y-lead(y))<1.5 & abs(z-lead(z))<1.5))%>% #filter out jumps of 1.5m and more
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
  mutate(x_approx = na.approx(x,  na.rm = TRUE),y_approx = na.approx(y,  na.rm = TRUE), z_approx = na.approx(z,  na.rm = TRUE))%>% #first linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 40cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # second linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 40cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # third linear interpolation
  filter((abs(x_approx-lag(x_approx))<0.4 & abs(y_approx-lag(y_approx))<0.4 & abs(z_approx-lag(z_approx))<0.4 & abs(x_approx-lead(x_approx))<0.4 & abs(y_approx-lead(y_approx))<0.4 & abs(z_approx-lead(z_approx))<0.4))%>% #filter out jumps of 40cm and more between two consecutive frames
  complete(frame_count=seq(from=min(frame_count, na.rm=TRUE), to=max(frame_count, na.rm=TRUE)))%>% #add all missing frames
    mutate(x_approx = na.approx(x_approx, na.rm = TRUE),y_approx = na.approx(y_approx,  na.rm = TRUE), z_approx = na.approx(z_approx,  na.rm = TRUE))%>% # 4th linear interpolation
  mutate(x_approx_roll = zoo::rollmean(x_approx, k = 3, fill = NA), y_approx_roll = zoo::rollmean(y_approx, k = 3, fill = NA),z_approx_roll = zoo::rollmean(z_approx, k = 3, fill = NA))%>% #rolling average
  mutate(x1=lead(x_approx_roll), y1=lead(y_approx_roll), z1=lead(z_approx_roll))%>% #add xyz coordinates of next frame
  mutate(distance=sqrt((x1-x_approx_roll)^2+(y1-y_approx_roll)^2+(z1-z_approx_roll)^2))%>%
  mutate(distance_chair_door= sqrt((object.center.data$x[object.center.data$id=="chair_door"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="chair_door"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="chair_door"]-z_approx_roll)^2))%>%
  mutate(distance_chair_window= sqrt((object.center.data$x[object.center.data$id=="chair_window"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="chair_window"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="chair_window"]-z_approx_roll)^2))%>%
    mutate(distance_obj_dl= sqrt((object.center.data$x[object.center.data$id=="obj_dl"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="obj_dl"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="obj_dl"]-z_approx_roll)^2))%>%
      mutate(distance_obj_dr= sqrt((object.center.data$x[object.center.data$id=="obj_dr"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="obj_dr"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="obj_dr"]-z_approx_roll)^2))%>%
      mutate(distance_obj_wl= sqrt((object.center.data$x[object.center.data$id=="obj_wl"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="obj_wl"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="obj_wl"]-z_approx_roll)^2))%>%
      mutate(distance_obj_wr= sqrt((object.center.data$x[object.center.data$id=="obj_wr"]-x_approx_roll)^2+(object.center.data$y[object.center.data$id=="obj_wr"]-y_approx_roll)^2+(object.center.data$z[object.center.data$id=="obj_wr"]-z_approx_roll)^2))%>%
  mutate(duration_chair_window=as.numeric(ifelse(x_approx_roll>= chair_window_ia_xl & x_approx_roll<= chair_window_ia_xr & y_approx_roll>= chair_window_ia_yl & y_approx_roll<= chair_window_ia_yh, 1, 0)))%>%
    mutate(duration_chair_door=as.numeric(ifelse(x_approx_roll>= chair_door_ia_xl & x_approx_roll<= chair_door_ia_xr & y_approx_roll>= chair_door_ia_yl & y_approx_roll<= chair_door_ia_yh, 1, 0)))%>%
    mutate(duration_door_l=as.numeric(ifelse(x_approx_roll<= door_l_ia_xl & x_approx_roll>= door_l_ia_xr & y_approx_roll>= door_l_ia_yl & y_approx_roll<= door_l_ia_yh, 1, 0)))%>%
    mutate(duration_door_r=as.numeric(ifelse(x_approx_roll<= door_r_ia_xl & x_approx_roll>= door_r_ia_xr & y_approx_roll>= door_r_ia_yl & y_approx_roll<= door_r_ia_yh, 1, 0)))%>%
  inner_join(demographics.data)
```


### Calculate the area covered by the dog (based on grid with cells of 10 x 10 cm)
```{r}
vis.data<-xdata.int%>%
  select(subject, name, condition, frame_count, head_centre.snout, x_approx_roll, y_approx_roll, z_approx_roll)%>%
mutate(subject_condition=paste(subject, condition, sep="_"))%>%
  pivot_wider(names_from=name, values_from=c(x_approx_roll, y_approx_roll, z_approx_roll))%>%
  filter(!is.na(x_approx_roll_head_centre), !is.na(y_approx_roll_head_centre))
  


##area covered:
##all cells must be of the exact same size
##set approximate edge length of cells in m:
edge.len=0.50
##set limits of the area:
xlim=c(-3.5, 3.8)
ylim=c(-2.7, 3.7)

##source function needed:
source("./functions/area_functions.r")
##determine resolution of the grid in x- and y-direction:
nx=round(diff(xlim)/edge.len)
ny=round(diff(ylim)/edge.len)

xdata<-xdata%>%mutate(subject_condition=paste(subject, condition, sep="_"))
subjects = unique(xdata$subject)
names = unique(xdata$name)
conditions =unique(xdata$condition)
subject_conditions =unique(xdata$subject_condition)

sel = list()
sel2 = list()
ydata = list()
xx = list()
area_covered<-c()
for(subject_condition_ in subject_conditions){

##create vector with the positions used (needed because the first and last position are NA):
ydata[[subject_condition_]]<-vis.data%>%filter(subject_condition==subject_condition_)
sel[[subject_condition_]]=which(apply(!is.na(ydata[[subject_condition_]][, c("x_approx_roll_head_centre", "y_approx_roll_head_centre")]), 1, all))
sel2[[subject_condition_]]=1+which(sel[[subject_condition_]][-1]&sel[[subject_condition_]][-length(sel[[subject_condition_]])])
##determined cells covered:
xx[[subject_condition_]]=lapply(sel2[[subject_condition_]][-1], function(x){
	y=effort.per.cell(grid.left=xlim[1], grid.right=xlim[2], grid.bottom=ylim[1], grid.top=ylim[2], nx=nx, ny=ny, 
		xs=ydata[[subject_condition_]]$x_approx_roll_head_centre[x-1], xe=ydata[[subject_condition_]]$x_approx_roll_head_centre[x], 
		ys=ydata[[subject_condition_]]$y_approx_roll_head_centre[x-1], ye=ydata[[subject_condition_]]$y_approx_roll_head_centre[x])
	apply(y[["effort"]][, c("x.name", "y.name")], 1, paste, collapse="_")
})
##and the area covered is:
area_covered<-c(area_covered, length(unique(unlist(xx[[subject_condition_]])))/(nx*ny))
}

area.data<-as.data.frame(cbind(subject_conditions, area_covered))%>%
  separate(subject_conditions, c("subject", "condition"))%>%
  mutate(name="head_centre", area_covered=as.numeric(area_covered))

area_covered_50cm<-ggplot(data=area.data, aes(x=condition, y=area_covered))+
  geom_boxplot()+
  geom_point()+
  theme_classic()
area_covered_50cm
```

### calculate visual angle
```{r}
source("./functions/angle_functions.r")
eye.frac=0.5
eye.frac.angle=pi/4 #45 degrees

vis.data<-xdata.int%>%
  select(subject, name, condition, frame_count, head_centre.snout, x_approx_roll, y_approx_roll, z_approx_roll)%>%
mutate(subject_condition=paste(subject, condition, sep="_"))%>%
  pivot_wider(names_from=name, values_from=c(x_approx_roll, y_approx_roll, z_approx_roll))%>%
    mutate(snout_hc_distance = sqrt((x_approx_roll_head_centre - x_approx_roll_snout)^2 + (y_approx_roll_head_centre - y_approx_roll_snout)^2 + (z_approx_roll_head_centre - z_approx_roll_snout)^2))


##determine all viewing angles for a track:
##object comprises is the location of the object to which you want to determine the viewing angles:
chair_door_loc=c(x=0.01, y=3.255, z=0.5)
chair_window_loc=c(x=0.01, y=-2.175, z=0.5)
chair_door_angles=sapply(1:nrow(vis.data), function(x){

	if(all(!is.na(vis.data[x, c("x_approx_roll_head_centre", "y_approx_roll_head_centre", "z_approx_roll_head_centre", 
		"x_approx_roll_snout", "y_approx_roll_snout", "z_approx_roll_snout")]))){
		xx<-dog.view.angle.and.eye.pos(head.pos=c(x=vis.data$x_approx_roll_head_centre[x],	
		                                        y=vis.data$y_approx_roll_head_centre[x],						                                        z=vis.data$z_approx_roll_head_centre[x]),
														snout.pos=c(x=vis.data$x_approx_roll_snout[x], 
																				y=vis.data$y_approx_roll_snout[x], 
																				z=vis.data$z_approx_roll_snout[x]), 
														eye.frac=eye.frac, eye.angle=eye.frac.angle)
		return(as.vector(v.angle(eye.loc=xx$eye.pos, object=chair_door_loc, eye.angle=xx$angle)))
	}else{
		return(NA)
	}
})

chair_window_angles=sapply(1:nrow(vis.data), function(x){
	if(all(!is.na(vis.data[x, c("x_approx_roll_head_centre", "y_approx_roll_head_centre", "z_approx_roll_head_centre", 
		"x_approx_roll_snout", "y_approx_roll_snout", "z_approx_roll_snout")]))){
		xx<-dog.view.angle.and.eye.pos(head.pos=c(x=vis.data$x_approx_roll_head_centre[x],	
		                                        y=vis.data$y_approx_roll_head_centre[x],						                                        z=vis.data$z_approx_roll_head_centre[x]),
														snout.pos=c(x=vis.data$x_approx_roll_snout[x], 
																				y=vis.data$y_approx_roll_snout[x], 
																				z=vis.data$z_approx_roll_snout[x]), 
														eye.frac=eye.frac, eye.angle=eye.frac.angle)
		return(as.vector(v.angle(eye.loc=xx$eye.pos, object=chair_window_loc, eye.angle=xx$angle)))
	}else{
		return(NA)
	}
})

vis.angle.data<-as.data.frame(cbind(vis.data, vis_angle_chair_door=as.numeric(chair_door_angles), vis_angle_chair_window=as.numeric(chair_window_angles)))

vis_angle=pi/6

vis.angle.summary.data<-vis.angle.data%>%
  group_by(subject, condition)%>%
  summarise(prop_look_chair_door=sum(vis_angle_chair_door<(vis_angle), na.rm = TRUE)/length(vis_angle_chair_door), mean_angle_chair_door=mean(vis_angle_chair_door, na.rm = TRUE), min_angle_chair_door=min(vis_angle_chair_door, na.rm = TRUE), prop_look_chair_window=sum(vis_angle_chair_window<(vis_angle), na.rm = TRUE)/length(vis_angle_chair_window), mean_angle_chair_window=mean(vis_angle_chair_window, na.rm = TRUE), min_angle_chair_window=min(vis_angle_chair_window, na.rm = TRUE))
```
### Tail angle
First the proportion of data available for the tail angle
```{r}
source("./functions/tail_angle_function.r")

frame.count.df<- xdata.int %>%
  ungroup()%>%
  group_by(subject,condition,  frame_count)%>%
  summarise(frame_count=mean(frame_count))%>%
  ungroup()

tailtip.data.raw<-xdata.int%>%
  ungroup()%>%
  select(subject, name, condition, frame_count, x, y)%>%
  filter(name=="tail_tip")%>%
  full_join(frame.count.df)%>%
  mutate(name="tail_tip")

atlas.data.raw<-xdata.int%>%
  ungroup()%>%
  select(subject, name, condition,frame_count, x, y)%>%
  filter(name=="atlas")%>%
  full_join(frame.count.df)%>%
  mutate(name="atlas")

tailbase.data.raw<-xdata.int%>%
  ungroup()%>%
  select(subject, name, condition, frame_count, x, y)%>%
  filter(name=="tail_base")%>%
  full_join(frame.count.df)%>%
  mutate(name="tail_base")



tail.data_raw<-tailtip.data.raw%>%
  full_join(atlas.data.raw)%>%
  full_join(tailbase.data.raw)%>%
  mutate(subject_condition=paste(subject, condition, sep="_"))%>%
  group_by(subject, condition, frame_count)%>%
  pivot_wider(names_from=name, values_from=c(x, y))%>%
  unchop(everything())


tail_angles_raw=sapply(1:nrow(tail.data_raw), function(x){
	if(all(!is.na(tail.data_raw[x, c("x_atlas", "y_atlas", "x_tail_base", "y_tail_base", "x_tail_tip", "y_tail_tip")]))){

		v.angle(tailtip.loc=c(x=tail.data_raw$x_tail_tip[x],	
		                                        y=tail.data_raw$y_tail_tip[x]),
		                   tailbase.loc=c(x=tail.data_raw$x_tail_base[x],	
		                                        y=tail.data_raw$y_tail_base[x]),
														atlas.loc=c(x=tail.data_raw$x_atlas[x], 
																				y=tail.data_raw$y_atlas[x]))
	}else{
		return(NA)
	}
})


tail.angle.data_raw<-as.data.frame(cbind(tail.data_raw, tail_angle=as.numeric(tail_angles_raw)))


```

```{r}
source("./functions/tail_angle_function.r")

tail.data<-xdata.int%>%
  select(subject, name, condition, frame_count, x_approx_roll, y_approx_roll)%>%
mutate(subject_condition=paste(subject, condition, sep="_"))%>%
  pivot_wider(names_from=name, values_from=c(x_approx_roll, y_approx_roll))


tail_angles=sapply(1:nrow(tail.data), function(x){
	if(all(!is.na(tail.data[x, c("x_approx_roll_atlas", "y_approx_roll_atlas", "x_approx_roll_tail_base", "y_approx_roll_tail_base", "x_approx_roll_tail_tip", "y_approx_roll_tail_tip")]))){

		v.angle(tailtip.loc=c(x=tail.data$x_approx_roll_tail_tip[x],	
		                                        y=tail.data$y_approx_roll_tail_tip[x]),
		                   tailbase.loc=c(x=tail.data$x_approx_roll_tail_base[x],	
		                                        y=tail.data$y_approx_roll_tail_base[x]),
														atlas.loc=c(x=tail.data$x_approx_roll_atlas[x], 
																				y=tail.data$y_approx_roll_atlas[x]))
	}else{
		return(NA)
	}
})


tail.angle.data<-as.data.frame(cbind(tail.data, tail_angle=as.numeric(tail_angles)))
```

```{r}

tail.angle.ind.data.raw<-tail.angle.data_raw%>%
  group_by(subject, condition)%>%
  summarise(tail_angle_prop_tracked=1-(sum(is.na(tail_angle))/length(tail_angle)))%>%
  ungroup()

tail.angle.ind.data<-tail.angle.data%>%
  group_by(subject, condition)%>%
  summarise(tail_angle_mean=mean(tail_angle, na.rm=TRUE), tail_angle_median=median(tail_angle, na.rm=TRUE))%>%
  ungroup()%>%
  full_join(tail.angle.ind.data.raw)


tail_plot<- ggplot(data=tail.angle.ind.data, aes(x=reorder(subject, tail_angle_median), y=tail_angle_median))+
  geom_bar(stat="identity", alpha=tail.angle.ind.data$tail_angle_prop_tracked)+
  facet_wrap(~condition)+
 # theme(axis.text.x = element_text(angle = 90,  vjust=0))+
  theme_bw()+
  coord_flip()+
  ylab("Median tail angle")+
  xlab("Subjects")

ggsave(tail_plot, filename = "graphics/tail_plot.png", width = 12, height=12, scale=0.7)
```


```{r}
tail_ttest_data<-tail.angle.data%>%
  group_by(subject, condition)%>%
  summarise(median=median(tail_angle, na.rm=TRUE))%>%
  #filter(condition=="op"| condition=="bp")%>%
  pivot_wider(names_from=condition, values_from=median)

t.test(tail_ttest_data$op, tail_ttest_data$bp, paired=TRUE)
t.test(tail_ttest_data$op)
t.test(tail_ttest_data$sp)
t.test(tail_ttest_data$bp)
t.test(tail_ttest_data$np)




```



### summary distance data (for analysis)
```{r}
distance.data<-xdata.int%>%
  group_by(subject, name, condition, owner_chair, stranger_chair)%>% 
  summarise(distance.sum=sum(distance, na.rm=TRUE),
            max.frame.count=max(frame_count), tracked_data_count=sum(!is.na(x)), prop_tracked_data= tracked_data_count/max.frame.count, median_err=median(err, na.rm=TRUE),
            mean.distance.chair.door=mean(distance_chair_door, na.rm=TRUE), mean.distance.chair.window=mean(distance_chair_window, na.rm=TRUE), min.distance.chair.door=min(distance_chair_door, na.rm=TRUE), min.distance.chair.window=min(distance_chair_window, na.rm=TRUE), min.distance.obj.dr=min(distance_obj_dr, na.rm=TRUE), min.distance.obj.dl=min(distance_obj_dl, na.rm=TRUE),  min.distance.obj.wr=min(distance_obj_wr, na.rm=TRUE), min.distance.obj.wl=min(distance_obj_wl, na.rm=TRUE), prop_chair_door=sum(duration_chair_door, na.rm = TRUE)/length(duration_chair_door), prop_chair_window=sum(duration_chair_window, na.rm = TRUE)/length(duration_chair_window), prop_door_l=sum(duration_door_l, na.rm = TRUE)/length(duration_door_l), prop_door_r=sum(duration_door_r, na.rm = TRUE)/length(duration_door_r))%>%
  full_join(area.data)%>%
  full_join(vis.angle.summary.data)%>%
  full_join(tail.angle.ind.data) %>%
  mutate(min_distance_owner_chair=ifelse(owner_chair=="window",min.distance.chair.window,ifelse(owner_chair=="door", min.distance.chair.door, NA)), mean_distance_owner_chair=ifelse(owner_chair=="window",mean.distance.chair.window,ifelse(owner_chair=="door", mean.distance.chair.door, NA)), prop_look_owner_chair=ifelse(owner_chair=="window",prop_look_chair_window,ifelse(owner_chair=="door", prop_look_chair_door, NA)), prop_look_stranger_chair=ifelse(stranger_chair=="window",prop_look_chair_window,ifelse(stranger_chair=="door", prop_look_chair_door, NA)), mean_vis_angle_stranger=ifelse(stranger_chair=="window",mean_angle_chair_window,ifelse(stranger_chair=="door", mean_angle_chair_door, NA)), mean_vis_angle_owner=ifelse(owner_chair=="window",mean_angle_chair_window,ifelse(owner_chair=="door", mean_angle_chair_door, NA)), min_vis_angle_stranger=ifelse(stranger_chair=="window",min_angle_chair_window,ifelse(stranger_chair=="door", min_angle_chair_door, NA)), min_vis_angle_owner=ifelse(owner_chair=="window",min_angle_chair_window,ifelse(owner_chair=="door", min_angle_chair_door, NA)), prop_duration_owner=ifelse(owner_chair=="window",prop_chair_window,ifelse(owner_chair=="door", prop_chair_door, NA)), prop_duration_stranger=ifelse(stranger_chair=="window",prop_chair_window,ifelse(stranger_chair=="door", prop_chair_door, NA)), prop_duration_door=prop_door_l+prop_door_r)
  
summary(distance.data)
write.csv(distance.data, file = "data/distance_travelled_summary_data.csv")
```


Visualise proportion looking to owner
```{r}
ggplot(data=distance.data, aes(x=condition, y=prop_look_owner_chair))+
  geom_boxplot()+
    geom_violin()+
  geom_point()+
  theme_classic()


ggplot(data=distance.data, aes(x=condition, y=prop_look_stranger_chair))+
  geom_boxplot()+
  geom_point()+
  theme_classic()


ggplot(data=distance.data, aes(x=condition, y=mean_vis_angle_owner))+
  geom_boxplot()+
  geom_point()+
  theme_classic()

ggplot(data=distance.data, aes(x=condition, y=mean_vis_angle_stranger))+
  geom_boxplot()+
  geom_point()+
  theme_classic()

ggplot(data=distance.data, aes(x=condition, y=min_vis_angle_owner))+
  geom_boxplot()+
  geom_point()+
  theme_classic()

ggplot(data=distance.data, aes(x=condition, y=min_vis_angle_stranger))+
  geom_boxplot()+
  geom_point()+
  theme_classic()
```


### Proportion of tracked data 
```{r}
#proportion of tracked data by keypoint
prop_tracked_by_keypoint<-distance.data%>%
  group_by(name)%>%
  summarise(mean(prop_tracked_data),median(prop_tracked_data), min(prop_tracked_data),max(prop_tracked_data), mean(median_err))

write.csv(prop_tracked_by_keypoint, file = "saves/prop_tracked_by_keypoint.csv")

#proportion of tracked data by condition
prop_tracked_by_condition<-distance.data%>%
  group_by(condition)%>%
  summarise(mean(prop_tracked_data), min(prop_tracked_data),max(prop_tracked_data), mean(median_err))

write.csv(prop_tracked_by_condition, file = "saves/prop_tracked_by_condition.csv")

#data with less than 50% tracked
distance.data%>%
 filter(prop_tracked_data<.5)%>%
  arrange(name)
```


### Heatmap of raw data

```{r}
library(viridis)
p1<-ggplot(data = xdata.int%>%filter(name=="snout"), aes(x=x_approx_roll, y=y_approx_roll))+
  #stat_density2d(aes(fill=..level.., alpha=..level..), geom="polygon")+
  geom_density_2d_filled(aes(fill=..level.., alpha=..level..),contour_var = "ndensity")+#"count"
  facet_wrap(~condition)+
  theme_bw()+
  #scale_fill_viridis() +
  theme(legend.position = "none")
p1

ggsave(p1, filename = "graphics/density_plot_37_dogs.pdf", width = 9,height=7, scale=0.8 )

p2<-ggplot(data = xdata.int%>%filter(name=="snout"), aes(x=x_approx_roll, y=y_approx_roll))+
  stat_density2d(aes(fill = ..density..), geom = "tile", contour = FALSE)+
  facet_wrap(~condition)+
  theme_bw()
p2
```

```{r}


xdata.int<-xdata.int%>%   mutate(condition2=fct_recode(as.factor(condition), "O+ / S+"="bp", "O- / S+"="sp", "O+ / S-"="op", "O- / S-"="np"))%>%
  mutate(condition2=fct_relevel(condition2, "O+ / S+", "O+ / S-","O- / S+",  "O- / S-"))



p3<-ggplot(data = xdata.int%>%filter(name=="head_centre"), aes(x=x_approx_roll, y=y_approx_roll))+
  facet_wrap(~condition2)+
  theme_bw()+
  #scale_fill_viridis() +
  theme(legend.position = "none")+
  annotate("rect", xmin=object.data$x[object.data$id=="dl1"], xmax=object.data$x[object.data$id=="dl2"], ymin=object.data$y[object.data$id=="dl1"],ymax=object.data$y[object.data$id=="dl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="dr1"], xmax=object.data$x[object.data$id=="dr2"], ymin=object.data$y[object.data$id=="dr1"],ymax=object.data$y[object.data$id=="dr2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wl1"], xmax=object.data$x[object.data$id=="wl2"], ymin=object.data$y[object.data$id=="wl1"],ymax=object.data$y[object.data$id=="wl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wr1"], xmax=object.data$x[object.data$id=="wr2"], ymin=object.data$y[object.data$id=="wr1"],ymax=object.data$y[object.data$id=="wr2"], fill="darkgrey", alpha=0.6)+
        annotate("rect", xmin=object.data$x[object.data$id=="door1"], xmax=object.data$x[object.data$id=="door2"], ymin=object.data$y[object.data$id=="door1"],ymax=object.data$y[object.data$id=="door2"], fill="black", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="window1"], xmax=object.data$x[object.data$id=="window2"], ymin=object.data$y[object.data$id=="window1"],ymax=object.data$y[object.data$id=="window2"], fill="black", alpha=0.6)+
      annotate("rect", xmin=1.619, xmax=2.466, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#right door
        annotate("rect", xmin=-2.265, xmax=-1.418, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#left door
          annotate("rect", xmin=-2.265, xmax=-1.265, ymin=2.7,ymax=3.7, color="black", fill=NA, alpha=0.8)+#left aoi
      annotate("rect", xmin=1.466, xmax=2.466, ymin=2.7,ymax=3.7, color="black", fill=NA, alpha=0.8)+#left aoi
            annotate("rect", xmin=-0.99, xmax=1.01, ymin=-0.875,ymax=-2.175, color="black", fill=NA, alpha=0.8)+#chair window aoi
      annotate("rect", xmin=-0.99, xmax=1.01, ymin=1.955,ymax=3.255, color="black", fill=NA, alpha=0.8)+#chair door aoi
    geom_point(alpha=0.012, color="dodgerblue")+
    geom_density_2d(color="red")+
    theme_bw()+ 
    ylab("Y coordinates (in m)") + xlab("X coordinates (in m)") + 
    xlim(-3.5, 3.8) + ylim(-2.7, 4)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank(), axis.line = element_line(colour = "black"), text=element_text(size=20))


p3



ggsave(p3, filename = "graphics/density_plot_37_dogs2.png", width = 8,height=7, scale=0.7 )

```


Plot exemplary data
```{r}
key.labs <- c("Head centre","Tail tip")
names(key.labs) <- c("head_centre", "tail_tip")


p.example<-ggplot(data = xdata.int%>%filter((name=="head_centre"|name=="tail_tip") & subject=="dawn" &condition=="bp"), aes(x=x_approx_roll, y=y_approx_roll))+
  facet_wrap(~name, labeller=labeller(name=key.labs))+
  theme_bw()+
  #scale_fill_viridis() +
  theme(legend.position = "none")+
  annotate("rect", xmin=object.data$x[object.data$id=="dl1"], xmax=object.data$x[object.data$id=="dl2"], ymin=object.data$y[object.data$id=="dl1"],ymax=object.data$y[object.data$id=="dl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="dr1"], xmax=object.data$x[object.data$id=="dr2"], ymin=object.data$y[object.data$id=="dr1"],ymax=object.data$y[object.data$id=="dr2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wl1"], xmax=object.data$x[object.data$id=="wl2"], ymin=object.data$y[object.data$id=="wl1"],ymax=object.data$y[object.data$id=="wl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wr1"], xmax=object.data$x[object.data$id=="wr2"], ymin=object.data$y[object.data$id=="wr1"],ymax=object.data$y[object.data$id=="wr2"], fill="darkgrey", alpha=0.6)+
        annotate("rect", xmin=object.data$x[object.data$id=="door1"], xmax=object.data$x[object.data$id=="door2"], ymin=object.data$y[object.data$id=="door1"],ymax=object.data$y[object.data$id=="door2"], fill="black", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="window1"], xmax=object.data$x[object.data$id=="window2"], ymin=object.data$y[object.data$id=="window1"],ymax=object.data$y[object.data$id=="window2"], fill="black", alpha=0.6)+
annotate("rect", xmin=1.619, xmax=2.466, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#right door
annotate("rect", xmin=-2.265, xmax=-1.418, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#left door
    geom_path(alpha=0.2, color="dodgerblue") + 
    geom_point(alpha=0.05, color="dodgerblue")+
    theme_bw()+ 
    ylab("Y coordinates (in m)") + xlab("X coordinates (in m)") + 
    xlim(-3.5, 3.8) + ylim(-2.7, 4)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank(), axis.line = element_line(colour = "black"), text=element_text(size=14))


p.example



ggsave(p.example, filename = "graphics/examplary_plot_dawn_bp.png", width = 12,height=6, scale=0.7 )

```
addition of tail angle information
```{r}

p.example2.data_x<-tail.angle.data%>%filter( subject=="dawn" &condition=="bp")%>%
  select(subject, condition, frame_count, x_approx_roll_tail_tip, x_approx_roll_head_centre)%>%
  pivot_longer(c(x_approx_roll_tail_tip, x_approx_roll_head_centre), names_to = "name", values_to =  "x_approx_roll")%>%
  mutate(name=str_remove(name, "x_approx_roll_"))

p.example2.data<-tail.angle.data%>%filter( subject=="dawn" &condition=="bp")%>%
  select(subject, condition, frame_count, y_approx_roll_tail_tip, y_approx_roll_head_centre)%>%
  pivot_longer(c(y_approx_roll_tail_tip, y_approx_roll_head_centre), names_to = "name", values_to =  "y_approx_roll")%>%
  mutate(name=str_remove(name, "y_approx_roll_"))%>%
  full_join(p.example2.data_x)%>%
  inner_join(tail.angle.data)

p.example2<-ggplot(data = p.example2.data%>%filter((name=="head_centre"|name=="tail_tip") & subject=="dawn" &condition=="bp"), aes(x=x_approx_roll, y=y_approx_roll))+
  facet_wrap(~name, labeller=labeller(name=key.labs))+
  theme_bw()+
  #scale_fill_viridis() +
  theme(legend.position = "none")+
  annotate("rect", xmin=object.data$x[object.data$id=="dl1"], xmax=object.data$x[object.data$id=="dl2"], ymin=object.data$y[object.data$id=="dl1"],ymax=object.data$y[object.data$id=="dl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="dr1"], xmax=object.data$x[object.data$id=="dr2"], ymin=object.data$y[object.data$id=="dr1"],ymax=object.data$y[object.data$id=="dr2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wl1"], xmax=object.data$x[object.data$id=="wl2"], ymin=object.data$y[object.data$id=="wl1"],ymax=object.data$y[object.data$id=="wl2"], fill="darkgrey", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="wr1"], xmax=object.data$x[object.data$id=="wr2"], ymin=object.data$y[object.data$id=="wr1"],ymax=object.data$y[object.data$id=="wr2"], fill="darkgrey", alpha=0.6)+
        annotate("rect", xmin=object.data$x[object.data$id=="door1"], xmax=object.data$x[object.data$id=="door2"], ymin=object.data$y[object.data$id=="door1"],ymax=object.data$y[object.data$id=="door2"], fill="black", alpha=0.6)+
    annotate("rect", xmin=object.data$x[object.data$id=="window1"], xmax=object.data$x[object.data$id=="window2"], ymin=object.data$y[object.data$id=="window1"],ymax=object.data$y[object.data$id=="window2"], fill="black", alpha=0.6)+
annotate("rect", xmin=1.619, xmax=2.466, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#right door
annotate("rect", xmin=-2.265, xmax=-1.418, ymin=3.7,ymax=3.9, fill="darkgreen", alpha=0.8)+#left door
    geom_path(alpha=0.2) +
  geom_point(data=p.example2.data%>%filter((name=="head_centre") & subject=="dawn" &condition=="bp"), alpha=0.05, color="dodgerblue")+
  geom_point(data=p.example2.data%>%filter((name=="tail_tip") & subject=="dawn" &condition=="bp"), aes(color = tail_angle), alpha=0.15, size = 1) +#
  scale_color_gradient("Tail angle", low = "yellow", high = "darkblue", breaks=c(-90, 0, 90), labels=c("90 (left)", "0", "90 (right)")) + 
    theme_bw()+ 
    ylab("Y coordinates (in m)") + xlab("X coordinates (in m)") + 
    xlim(-3.5, 3.8) + ylim(-2.7, 4)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank(), axis.line = element_line(colour = "black"), text=element_text(size=14))


p.example2

```

## Head centre - snout distance: comparison between reconstructed and measured values
```{r}
hs.data<-vis.data%>%
  group_by(subject)%>%
  summarise(measured_hs_distance=mean(head_centre.snout)/100, reconstructed_hs_distance=median(snout_hc_distance, na.rm = TRUE))


hs_plot<-ggplot(data=vis.data, aes(x=snout_hc_distance))+
  geom_histogram()+
  facet_wrap(~subject)+
  geom_vline(data=hs.data, aes(xintercept=measured_hs_distance), color="red", lwd=1)+
    geom_vline(data=hs.data, aes(xintercept=reconstructed_hs_distance), color="blue",lty=2, lwd=1)+
  xlim(0,0.4)+
  theme_bw()


hs.data2<-hs.data%>%
  rename("Reconstructed (median)"="reconstructed_hs_distance", Measured="measured_hs_distance")%>%
  pivot_longer(names_to = "source", values_to="hs_dis", c(Measured, "Reconstructed (median)"))


hs_plot2<-ggplot(data=hs.data2, aes(x=source, y=hs_dis))+
  #geom_boxplot+
  geom_point(alpha=0.25)+
  geom_line(aes(group=subject), lty=2, alpha=0.4)+
  theme_bw()+
  ylab("Head centre - Snout distance (in m)")+
  xlab("")+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_blank(), axis.line = element_line(colour = "black"))
hs_plot2

ggsave(hs_plot, filename = "graphics/head_snot_distance_histogram.pdf", width=16, height=9, scale=0.8)

cor.test(hs.data$measured_hs_distance, hs.data$reconstructed_hs_distance, method="pearson")
```
